pyproject.toml
[project]
name = "llm-code-injection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=4.42.0",
    "flask>=3.0.3",
    "transformers>=4.44.2",
    "torch>=2.4.0",
    "crewai>=0.51.1",
    "crewai-tools>=0.8.3",
    "langchain-community>=0.2.12",
    "pynput>=1.7.7",
    "python-dotenv>=1.0.1",
    "pydantic>=2.8.2",
    "ipykernel>=6.29.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
